LayoutTests/media/modern-media-controls/time-label/time-label-white-space-nowrap.html

<!DOCTYPE html>
<script src="../../../resources/js-test-pre.js"></script>
<body>
<video src="../../content/test.mp4" style="position: absolute; left: 0; top: 0; width: 800px;" controls autoplay></video>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing that time labels inside media controls have 'white-space: nowrap'.");

let timeLabel;

document.querySelector("video").addEventListener("play", (event) => {
    const media = event.target;
    media.pause();

    window.requestAnimationFrame(() => {
        timeLabel = window.internals.shadowRoot(media).querySelector(".time-label");
        shouldBeEqualToString("timeLabel.ownerDocument.defaultView.getComputedStyle(timeLabel).whiteSpace", "nowrap");

        debug("");
        media.remove();
        finishJSTest();
    });
});

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
